APIListScheduleResponse now uses int type instead of Number type

+++ src/main/java/com/urbanairship/api/client/parse/APIListScheduleResponseReader.java
@@ -18,8 +18,8 @@ public class APIListScheduleResponseReader implements JsonObjectReader<APIListSc
 
     public APIListScheduleResponseReader() { this.builder = APIListScheduleResponse.newBuilder(); }
 
-    public void readCount(JsonParser jsonParser) throws IOException { builder.setCount(jsonParser.readValueAs(Number.class)); }
-    public void readTotalCount(JsonParser jsonParser) throws IOException { builder.setTotalCount(jsonParser.readValueAs(Number.class)); }
+    public void readCount(JsonParser jsonParser) throws IOException { builder.setCount(jsonParser.readValueAs(Number.class).intValue()); }
+    public void readTotalCount(JsonParser jsonParser) throws IOException { builder.setTotalCount(jsonParser.readValueAs(Number.class).intValue()); }
     public void readListScheduleResponse(JsonParser jsonParser) throws IOException { builder.setSchedule((List<ScheduleResponseObject>) jsonParser.readValueAs(new TypeReference<List<ScheduleResponseObject>>() {})); }
 
     @Override
